- Se crea ProductListViewModel, ProductListCell, ProductListScreen - Se crean modulos de inyecciôn de dependencias de ViewModel y de CoroutineScope - Se crea DtoMapper

+++ app/src/main/java/com/example/myapplication/domain/entities/data/Records.kt
@@ -4,8 +4,8 @@ package com.example.myapplication.domain.entities.data
 
 data class Records (
     var name: String ,
-    var listPrice : Int,
-    var promoPrice: Int,
+    var listPrice : Double,
+    var promoPrice: Double,
     var image: String,
     var colors: List<Colors>
 )
